mvp-services/requirements.txt
# Upbit LLM Analytics Platform - Core Dependencies

# Kafka 메시지 브로커
confluent-kafka==2.8.0

# PostgreSQL/TimescaleDB 연결
psycopg2-binary==2.9.10

# WebSocket 연결 (Upbit API)
websockets==14.2

# 환경 변수 관리
python-dotenv==1.0.1

# HTTP 요청 (필요시)
requests==2.32.3

# 기본 유틸리티
python-dateutil==2.9.0.post0

# Week 2 MCP + LLM 연동
openai==1.56.2

# 수치 계산 (이상 탐지용)
numpy==1.24.3

# Week 3-4에서 추가될 의존성들 (웹 인터페이스)
# 추후 추가 예정:
# - fastapi
# - uvicorn 
# - sqlalchemy (필요시)
